# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Snap API",
    "description": "Screenshot styling API powered by Microlink. Create beautiful, styled screenshots with custom borders, gradients, browser mockups, and more - similar to ray.so but for websites.",
    "version": "1.0.0",
    "contact": {
      "name": "Snap API Support",
      "url": "https://github.com/your-repo/snap"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "tags": [
    {
      "name": "Screenshots",
      "description": "Generate styled screenshots"
    },
    {
      "name": "Preview",
      "description": "Visual preview of screenshots"
    },
    {
      "name": "Health",
      "description": "API health and status"
    }
  ],
  "paths": {
    "/api": {
      "get": {
        "tags": ["Health"],
        "summary": "API Information",
        "description": "Returns basic information about the Snap API including available endpoints",
        "responses": {
          "200": {
            "description": "API information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiInfo"
                }
              }
            }
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "tags": ["Health"],
        "summary": "Health Check",
        "description": "Check if the API is running and healthy",
        "responses": {
          "200": {
            "description": "API is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/snap": {
      "get": {
        "tags": ["Screenshots"],
        "summary": "Generate Screenshot (Query Parameters)",
        "description": "Generate a styled screenshot using URL query parameters",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "description": "The URL to take a screenshot of",
            "schema": {
              "type": "string",
              "format": "uri",
              "example": "https://example.com"
            }
          },
          {
            "name": "borderRadius",
            "in": "query",
            "description": "Border radius in pixels (0-50)",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 50,
              "example": 15
            }
          },
          {
            "name": "margin",
            "in": "query",
            "description": "Margin around the image in pixels (0-200)",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 200,
              "example": 40
            }
          },
          {
            "name": "background.type",
            "in": "query",
            "description": "Background type",
            "schema": {
              "type": "string",
              "enum": ["solid", "gradient"],
              "example": "gradient"
            }
          },
          {
            "name": "background.color",
            "in": "query",
            "description": "Background color (for solid type)",
            "schema": {
              "type": "string",
              "pattern": "^#[0-9A-Fa-f]{6}$|^#[0-9A-Fa-f]{3}$",
              "example": "#ffffff"
            }
          },
          {
            "name": "background.direction",
            "in": "query",
            "description": "Gradient direction (for gradient type)",
            "schema": {
              "type": "string",
              "enum": ["to-r", "to-l", "to-t", "to-b", "to-br", "to-bl", "to-tr", "to-tl"],
              "example": "to-br"
            }
          },
          {
            "name": "browserMockup",
            "in": "query",
            "description": "Browser mockup style",
            "schema": {
              "type": "string",
              "enum": ["safari", "chrome", "firefox", "edge", "none"],
              "example": "safari"
            }
          },
          {
            "name": "viewport.width",
            "in": "query",
            "description": "Viewport width in pixels (320-4096)",
            "schema": {
              "type": "integer",
              "minimum": 320,
              "maximum": 4096,
              "example": 1920
            }
          },
          {
            "name": "viewport.height",
            "in": "query",
            "description": "Viewport height in pixels (240-4096)",
            "schema": {
              "type": "integer",
              "minimum": 240,
              "maximum": 4096,
              "example": 1080
            }
          },
          {
            "name": "colorScheme",
            "in": "query",
            "description": "Page color scheme",
            "schema": {
              "type": "string",
              "enum": ["light", "dark"],
              "example": "dark"
            }
          },
          {
            "name": "delay",
            "in": "query",
            "description": "Additional delay before screenshot in milliseconds (0-10000)",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 10000,
              "example": 2000
            }
          },
          {
            "name": "shadow.enabled",
            "in": "query",
            "description": "Enable drop shadow",
            "schema": {
              "type": "boolean",
              "example": true
            }
          },
          {
            "name": "shadow.blur",
            "in": "query",
            "description": "Shadow blur radius (0-50)",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 50,
              "example": 20
            }
          },
          {
            "name": "finalSize.width",
            "in": "query",
            "description": "Final image width in pixels (100-4096)",
            "schema": {
              "type": "integer",
              "minimum": 100,
              "maximum": 4096,
              "example": 1200
            }
          },
          {
            "name": "finalSize.height",
            "in": "query",
            "description": "Final image height in pixels (100-4096)",
            "schema": {
              "type": "integer",
              "minimum": 100,
              "maximum": 4096,
              "example": 800
            }
          },
          {
            "name": "finalSize.aspectRatio",
            "in": "query",
            "description": "Final image aspect ratio (e.g., 16:9, 4:3, 1:1)",
            "schema": {
              "type": "string",
              "pattern": "^\\d+:\\d+$",
              "example": "16:9"
            }
          },
          {
            "name": "viewport.isMobile",
            "in": "query",
            "description": "Simulate mobile device for screenshot",
            "schema": {
              "type": "boolean",
              "example": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Screenshot generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Screenshot service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Screenshots"],
        "summary": "Generate Screenshot (JSON Body)",
        "description": "Generate a styled screenshot using a JSON request body with full configuration options",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapRequest"
              },
              "examples": {
                "simple": {
                  "summary": "Simple screenshot",
                  "value": {
                    "url": "https://example.com"
                  }
                },
                "styled": {
                  "summary": "Styled screenshot with gradient",
                  "value": {
                    "url": "https://github.com",
                    "style": {
                      "borderRadius": 15,
                      "margin": 40,
                      "background": {
                        "type": "gradient",
                        "direction": "to-br",
                        "colors": [
                          {"color": "#667eea", "position": 0},
                          {"color": "#764ba2", "position": 100}
                        ]
                      },
                      "browserMockup": "safari",
                      "shadow": {
                        "enabled": true,
                        "blur": 20,
                        "offsetY": 10
                      }
                    }
                  }
                },
                "advanced": {
                  "summary": "Advanced configuration",
                  "value": {
                    "url": "https://tailwindcss.com",
                    "screenshot": {
                      "viewport": {
                        "width": 1280,
                        "height": 720,
                        "isMobile": false
                      },
                      "colorScheme": "dark",
                      "delay": 3000
                    },
                    "style": {
                      "borderRadius": 12,
                      "margin": 60,
                      "background": {
                        "type": "gradient",
                        "direction": "to-r",
                        "colors": [
                          {"color": "#FF6B6B", "position": 0},
                          {"color": "#4ECDC4", "position": 50},
                          {"color": "#45B7D1", "position": 100}
                        ]
                      },
                      "browserMockup": "chrome",
                      "finalSize": {
                        "width": 1200,
                        "aspectRatio": "16:9",
                        "maintainAspectRatio": true
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Screenshot generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Screenshot service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/preview": {
      "get": {
        "tags": ["Preview"],
        "summary": "Preview Screenshot in Browser",
        "description": "Generate and display a screenshot in an HTML page with download options",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "description": "The URL to take a screenshot of",
            "schema": {
              "type": "string",
              "format": "uri",
              "example": "https://example.com"
            }
          },
          {
            "name": "borderRadius",
            "in": "query",
            "description": "Border radius in pixels",
            "schema": {
              "type": "integer",
              "example": 15
            }
          },
          {
            "name": "margin",
            "in": "query",
            "description": "Margin in pixels",
            "schema": {
              "type": "integer",
              "example": 40
            }
          }
        ],
        "responses": {
          "200": {
            "description": "HTML page with screenshot preview",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "example": "<!DOCTYPE html><html>...</html>"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid URL parameter",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Snap API"
          },
          "version": {
            "type": "string",
            "example": "1.0.0"
          },
          "description": {
            "type": "string",
            "example": "Screenshot styling API powered by Microlink"
          },
          "endpoints": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2023-12-07T10:30:00.000Z"
          }
        }
      },
      "SnapRequest": {
        "type": "object",
        "required": ["url"],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "description": "The URL to take a screenshot of",
            "example": "https://example.com"
          },
          "screenshot": {
            "$ref": "#/components/schemas/ScreenshotConfig"
          },
          "style": {
            "$ref": "#/components/schemas/StyleConfig"
          }
        }
      },
      "ScreenshotConfig": {
        "type": "object",
        "properties": {
          "waitUntil": {
            "type": "string",
            "enum": ["load", "domcontentloaded", "networkidle0", "networkidle2"],
            "default": "networkidle0",
            "description": "When to consider navigation complete"
          },
          "delay": {
            "type": "integer",
            "minimum": 0,
            "maximum": 10000,
            "default": 2000,
            "description": "Additional delay in milliseconds"
          },
          "colorScheme": {
            "type": "string",
            "enum": ["light", "dark"],
            "default": "dark",
            "description": "Page color scheme preference"
          },
          "viewport": {
            "$ref": "#/components/schemas/ViewportConfig"
          }
        }
      },
      "ViewportConfig": {
        "type": "object",
        "properties": {
          "width": {
            "type": "integer",
            "minimum": 320,
            "maximum": 4096,
            "default": 1920,
            "description": "Viewport width in pixels"
          },
          "height": {
            "type": "integer",
            "minimum": 240,
            "maximum": 4096,
            "default": 1080,
            "description": "Viewport height in pixels"
          },
          "isMobile": {
            "type": "boolean",
            "default": false,
            "description": "Simulate mobile device"
          },
          "deviceScaleFactor": {
            "type": "number",
            "minimum": 0.5,
            "maximum": 3,
            "default": 1,
            "description": "Device pixel ratio"
          }
        }
      },
      "StyleConfig": {
        "type": "object",
        "properties": {
          "borderRadius": {
            "type": "integer",
            "minimum": 0,
            "maximum": 50,
            "default": 8,
            "description": "Border radius in pixels"
          },
          "margin": {
            "type": "integer",
            "minimum": 0,
            "maximum": 200,
            "default": 32,
            "description": "Margin around image in pixels"
          },
          "background": {
            "oneOf": [
              {"$ref": "#/components/schemas/SolidBackground"},
              {"$ref": "#/components/schemas/GradientBackground"}
            ]
          },
          "browserMockup": {
            "type": "string",
            "enum": ["safari", "chrome", "firefox", "edge", "none"],
            "default": "none",
            "description": "Browser mockup style"
          },
          "shadow": {
            "$ref": "#/components/schemas/ShadowConfig"
          },
          "finalSize": {
            "$ref": "#/components/schemas/FinalImageSize"
          }
        }
      },
      "SolidBackground": {
        "type": "object",
        "required": ["type", "color"],
        "properties": {
          "type": {
            "type": "string",
            "enum": ["solid"]
          },
          "color": {
            "type": "string",
            "pattern": "^#[0-9A-Fa-f]{6}$|^#[0-9A-Fa-f]{3}$",
            "description": "Hex color code",
            "example": "#ffffff"
          }
        }
      },
      "GradientBackground": {
        "type": "object",
        "required": ["type", "direction", "colors"],
        "properties": {
          "type": {
            "type": "string",
            "enum": ["gradient"]
          },
          "direction": {
            "type": "string",
            "enum": ["to-r", "to-l", "to-t", "to-b", "to-br", "to-bl", "to-tr", "to-tl"],
            "description": "Gradient direction"
          },
          "colors": {
            "type": "array",
            "minItems": 2,
            "maxItems": 5,
            "items": {
              "$ref": "#/components/schemas/GradientColor"
            }
          }
        }
      },
      "GradientColor": {
        "type": "object",
        "required": ["color"],
        "properties": {
          "color": {
            "type": "string",
            "pattern": "^#[0-9A-Fa-f]{6}$|^#[0-9A-Fa-f]{3}$",
            "description": "Hex color code",
            "example": "#667eea"
          },
          "position": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "description": "Position in gradient (0-100%)",
            "example": 50
          }
        }
      },
      "ShadowConfig": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "default": true,
            "description": "Enable drop shadow"
          },
          "blur": {
            "type": "integer",
            "minimum": 0,
            "maximum": 50,
            "default": 20,
            "description": "Shadow blur radius"
          },
          "offsetX": {
            "type": "integer",
            "minimum": -50,
            "maximum": 50,
            "default": 0,
            "description": "Horizontal shadow offset"
          },
          "offsetY": {
            "type": "integer",
            "minimum": -50,
            "maximum": 50,
            "default": 10,
            "description": "Vertical shadow offset"
          },
          "color": {
            "type": "string",
            "pattern": "^#[0-9A-Fa-f]{6}$|^#[0-9A-Fa-f]{3}$",
            "default": "#000000",
            "description": "Shadow color"
          },
          "opacity": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "default": 0.15,
            "description": "Shadow opacity"
          }
        }
      },
      "SnapResponse": {
        "type": "object",
        "required": ["success"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "imageUrl": {
            "type": "string",
            "description": "Base64 encoded image data URL",
            "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAA..."
          },
          "metadata": {
            "$ref": "#/components/schemas/ResponseMetadata"
          },
          "error": {
            "type": "string",
            "description": "Error message (only present when success is false)"
          }
        }
      },
      "ResponseMetadata": {
        "type": "object",
        "properties": {
          "originalUrl": {
            "type": "string",
            "format": "uri",
            "example": "https://example.com"
          },
          "processedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-12-07T10:30:00.000Z"
          },
          "dimensions": {
            "type": "object",
            "properties": {
              "width": {
                "type": "integer",
                "example": 1984
              },
              "height": {
                "type": "integer",
                "example": 1144
              }
            }
          }
        }
      },
      "FinalImageSize": {
        "type": "object",
        "properties": {
          "width": {
            "type": "integer",
            "minimum": 100,
            "maximum": 4096,
            "description": "Final image width in pixels"
          },
          "height": {
            "type": "integer",
            "minimum": 100,
            "maximum": 4096,
            "description": "Final image height in pixels"
          },
          "aspectRatio": {
            "type": "string",
            "pattern": "^\\d+:\\d+$",
            "description": "Aspect ratio (e.g., 16:9, 4:3, 1:1)",
            "example": "16:9"
          },
          "maintainAspectRatio": {
            "type": "boolean",
            "default": true,
            "description": "Whether to maintain aspect ratio when resizing"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["success", "error"],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Validation error: Invalid URL format"
          }
        }
      }
    }
  }
}